Refactor - extract common test helpers
namely XENOPS_CLI_CONTENTS and make_ramfs

related to xenserver/planex#53
